% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchull-package.R
\docType{package}
\name{cchull-package}
\alias{cchull}
\alias{cchull-package}
\title{cchull: Concave Hull}
\description{
Provide 'cpp' versions of concave hull algorithm using the concaveman library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/cchull}
  \item Report bugs at \url{https://github.com/hypertidy/cchull/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@email.com} (forked from original mpadge/concavetest)

Authors:
\itemize{
  \item Mark Padgham
}

Other contributors:
\itemize{
  \item Stanislaw Adaszewski (author of include concaveman-cpp code) [copyright holder]
}

}
\keyword{internal}
