% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concave.R
\name{cchull}
\alias{cchull}
\title{cchull concave hull}
\usage{
cchull(x, y, concavity = 2, length_threshold = NULL)
}
\arguments{
\item{x}{x values}

\item{y}{y values}
}
\value{
hull
}
\description{
cchull concave hull
}
